refactor: organize ui/ into domain subdirectories + close OpenAPI gaps
UI refactoring:
- Move screens into ui/onboarding/, ui/credentials/, ui/verification/
  matching domain boundaries (was 15 individual-file code-paths)
- Update package declarations and imports
- Simplify domains.yaml code-paths to directory-level entries

OpenAPI completeness:
- Add POST /sessions (verifier session creation)
- Add GET /.well-known/did.json (verifier DID document)
- Add GET /status/{listId} (StatusList2021 bitstring)
- Add POST /status/{listId}/revoke (credential revocation)
- Add GET /.well-known/jwks.json (issuer JWKS)
- Add schemas: SessionRequest, SessionResponse, StatusList2021Credential,
  RevokeRequest, JWKS, JWK
