+++ mobile/androidApp/src/main/kotlin/id/cachet/wallet/android/ui/WalletApp.kt
@@ -16,9 +16,19 @@ import id.cachet.wallet.android.ui.components.CachetSegmentedControl
 import id.cachet.wallet.android.ui.fixtures.DemoFixtures
 import id.cachet.wallet.android.ui.fixtures.HappyPathScenario
 import id.cachet.wallet.android.ui.fixtures.ScenarioRegistry
+import id.cachet.wallet.android.ui.credentials.ActivityScreen
+import id.cachet.wallet.android.ui.credentials.CachetDetailScreen
+import id.cachet.wallet.android.ui.credentials.HomeScreen
 import id.cachet.wallet.android.ui.mapper.CachPackMapper
 import id.cachet.wallet.android.ui.model.*
+import id.cachet.wallet.android.ui.onboarding.OnboardingScreen
 import id.cachet.wallet.android.ui.theme.*
+import id.cachet.wallet.android.ui.verification.CachetResultScreen
+import id.cachet.wallet.android.ui.verification.IncomingRequestScreen
+import id.cachet.wallet.android.ui.verification.PackPickerMode
+import id.cachet.wallet.android.ui.verification.PackPickerScreen
+import id.cachet.wallet.android.ui.verification.QrScannerScreen
+import id.cachet.wallet.android.ui.verification.QrShareScreen
 import kotlinx.serialization.json.*
 import org.koin.androidx.compose.koinViewModel
 import org.koin.core.parameter.parametersOf
