+++ mobile/androidApp/src/main/kotlin/id/cachet/wallet/android/ui/credentials/CachetDetailScreen.kt
@@ -1,4 +1,4 @@
-package id.cachet.wallet.android.ui
+package id.cachet.wallet.android.ui.credentials
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
