+++ mobile/androidApp/src/main/kotlin/id/cachet/wallet/android/ui/verification/IncomingRequestScreen.kt
@@ -1,4 +1,4 @@
-package id.cachet.wallet.android.ui
+package id.cachet.wallet.android.ui.verification
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
