+++ mobile/androidApp/src/main/kotlin/id/cachet/wallet/android/ui/verification/PackListScreen.kt
@@ -1,4 +1,4 @@
-package id.cachet.wallet.android.ui
+package id.cachet.wallet.android.ui.verification
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
