+++ mobile/androidApp/src/main/kotlin/id/cachet/wallet/android/ui/verification/QrShareScreen.kt
@@ -1,4 +1,4 @@
-package id.cachet.wallet.android.ui
+package id.cachet.wallet.android.ui.verification
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.*
